Refactor MPC to different file, change torque limits
